# OpenApi/MysiePysieService.yaml
openapi: 3.0.0
info:
  title: MysiePysie Service
  description: |
    Mysie Pysie Service for UZ Project with collaboration with Metapack
  contact:
    email: bartosz_musielak@hotmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
- url: https://virtserver.swaggerhub.com/MysiePysie/MysiePysie/0.0.1
  description: SwaggerHub API Auto Mocking
- url: localhost:8080
paths:
  /students:
    get:
      tags:
      - students
      summary: Returns all existing sutdents
      operationId: getStudents
      responses:
        "200":
          description: successful oepration
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        "404":
          description: Students not found
        "500":
          description: Server error
    put:
      tags:
      - students
      summary: Update an existing student
      operationId: updateStudent
      requestBody:
        $ref: '#/components/requestBodies/Student'
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Student not found
        "405":
          description: Validation exception
    post:
      tags:
      - students
      summary: Add a new student to the database
      operationId: addStudent
      requestBody:
        $ref: '#/components/requestBodies/Student'
      responses:
        "405":
          description: Invalid input
  /students/{studentId}:
    get:
      tags:
      - students
      summary: Find student by ID
      description: Returens a single student
      operationId: getStudentById
      parameters:
      - name: studentId
        in: path
        description: ID of student to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
            application/xml:
              schema:
                $ref: '#/components/schemas/Student'
        "400":
          description: Invalid ID supplied
        "404":
          description: Student not found
    post:
      tags:
      - students
      summary: Updates a student in the database with form data
      operationId: updateStudentWithForm
      parameters:
      - name: studentId
        in: path
        description: ID of student that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/body'
      responses:
        "405":
          description: Invalid input
    delete:
      tags:
      - students
      summary: Deletes a student
      operationId: deleteStudent
      parameters:
      - name: api_key
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: studentId
        in: path
        description: Student id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Student not found
...
  /subjects:
    get:
      tags:
      - subjects
      summary: Returns all existing subjects
      operationId: getSubjects
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
        "404":
          description: Subjects not found
        "500":
          description: Server error
    put:
      tags:
      - subjects
      summary: Update an existing subject
      operationId: updateSubject
      requestBody:
        $ref: '#/components/requestBodies/Subject'
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Subject not found
        "405":
          description: Validation exception
    post:
      tags:
      - subjects
      summary: Add a new subject to the database
      operationId: addSubject
      requestBody:
        $ref: '#/components/requestBodies/Subject'
      responses:
        "405":
          description: Invalid input
  /subjects/{subjectId}:
    get:
      tags:
      - subjects
      summary: Find subject by ID
      description: Returens a single subject
      operationId: getSubjectById
      parameters:
      - name: subjectId
        in: path
        description: ID of subject to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
            application/xml:
              schema:
                $ref: '#/components/schemas/Subject'
        "400":
          description: Invalid ID supplied
        "404":
          description: Subject not found
    post:
      tags:
      - subjects
      summary: Updates a subject in the database with form data
      operationId: updateSubjectWithForm
      parameters:
      - name: subjectId
        in: path
        description: ID of subject that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/body_2'
      responses:
        "405":
          description: Invalid input
    delete:
      tags:
      - subjects
      summary: Deletes a subject
      operationId: deleteSubject
      parameters:
      - name: subjectId
        in: path
        description: Subject id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Subject not found
  /classes:
    get:
      tags:
      - classes
      summary: Returns all existing classes
      operationId: getClasses
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Class'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Class'
        "404":
          description: Classes not found
        "500":
          description: Server error
    put:
      tags:
      - classes
      summary: Update an existing class
      operationId: updateClass
      requestBody:
        $ref: '#/components/requestBodies/Class'
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Class not found
        "405":
          description: Validation exception
    post:
      tags:
      - classes
      summary: Add a new class to the database
      operationId: addClass
      requestBody:
        $ref: '#/components/requestBodies/Class'
      responses:
        "405":
          description: Invalid input
  /classes/{classId}:
    get:
      tags:
      - classes
      summary: Find class by ID
      description: Returns a single class
      operationId: getClassById
      parameters:
      - name: classId
        in: path
        description: ID of class to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
            application/xml:
              schema:
                $ref: '#/components/schemas/Class'
        "400":
          description: Invalid ID supplied
        "404":
          description: Class not found
    post:
      tags:
      - classes
      summary: Updates a class in the database with form data
      operationId: updateClassWithForm
      parameters:
      - name: classId
        in: path
        description: ID of class that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/body_3'
      responses:
        "405":
          description: Invalid input
    delete:
      tags:
      - classes
      summary: Deletes a class
      operationId: deleteClass
      parameters:
      - name: classId
        in: path
        description: Class id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Class not found
...
components:
  schemas:
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int64
        forename:
          type: string
        surename:
          type: string
        age:
          type: integer
          format: int64
        status:
          type: string
      xml:
        name: Student
...
    Subject:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        ECTSpoints:
          type: integer
          format: int64
        teacher:
          type: object
      xml:
        name: Subject
    Class:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
      xml:
        name: Class
...
    body:
      type: object
      properties:
        forename:
          type: string
          description: Updated forename of a user
        surname:
          type: string
          description: Updated surname of a user
        class:
          type: string
          description: Updated class of a user
        age:
          type: integer
          description: Updated age of a user
          format: int64
        role:
          type: string
          description: Updated role of user
...
    body_2:
      type: object
      properties:
        name:
          type: string
          description: Updated name of a subject
        ECTSpoints:
          type: integer
          description: Updated ECTS point value of a subject
          format: int64
    body_3:
      type: object
      properties:
        name:
          type: string
          description: Updated name of a class
  requestBodies:
    Student:
      description: Student object that needs to be added to the database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Student'
        application/xml:
          schema:
            $ref: '#/components/schemas/Student'
      required: true
...
    Subject:
      description: Subject object that needs to be added to the database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Subject'
        application/xml:
          schema:
            $ref: '#/components/schemas/Subject'
      required: true
    Class:
      description: Class object that needs to be added to the database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Class'
        application/xml:
          schema:
            $ref: '#/components/schemas/Class'
      required: true
...